+++ CHANGELOG.md
@@ -1,5 +1,28 @@
-v0.9.2 - TBD
+v0.10.3 - TBD
+-------------
+
+v0.10.2 - 12/03/2019
+-------------
+
+v0.10.1 - 02/27/2019
+-------------
+
+### Fixed
+ - [Issue #153](https://github.com/uber/rides-android-sdk/issues/153) NullPointerException when login via SSO without setting product flow priority
+ - [Issue #151](https://github.com/uber/rides-android-sdk/issues/151) Login throws IllegalStateException when using only CustomScopes
+
+
+v0.10.0 - 12/14/2018
 ------------
+
+### Added
+ - [Issue #144](https://github.com/uber/rides-android-sdk/issues/144) Allow SSO Client to dictate which Uber Apps can be used for SSO
+ - [Issue #138](https://github.com/uber/rides-android-sdk/issues/138) Support for IETF RFC 8252
+ - [Issue #130](https://github.com/uber/rides-android-sdk/issues/130) Support for Uber Eats SSO
+
+### Fixed
+ - [Issue #129](https://github.com/uber/rides-android-sdk/issues/129) Allow use of refresh token for non-privileged scopes
+ - [Issue #119](https://github.com/uber/rides-android-sdk/issues/119) Redirect URL documentation issue
 
 v0.9.1 - 03/20/2018
 ------------